# api/docs/swagger.yaml
definitions:
...
  github_com_time_capsule_Api-Gateway-Timecapsule_models.CreateMemoryModel:
    properties:
      date:
        type: string
      description:
        type: string
      id:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      place_name:
        type: string
      privacy:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
      user_id:
        type: string
    type: object
...
  github_com_time_capsule_Api-Gateway-Timecapsule_models.PatchMemoryModel:
    properties:
      date:
        type: string
      description:
        type: string
      id:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      place_name:
        type: string
      privacy:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
    type: object
...
  github_com_time_capsule_Api-Gateway-Timecapsule_models.UpdateMemoryModel:
    properties:
      date:
        type: string
      description:
        type: string
      id:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      place_name:
        type: string
      privacy:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
      user_id:
        type: string
    type: object
...
  memory.GetAllMemoriesResponse:
    properties:
      count:
        type: integer
      memories:
        items:
          $ref: '#/definitions/memory.Memory'
        type: array
    type: object
...
  memory.Memory:
    properties:
      created_at:
        type: string
      date:
        type: string
      description:
        type: string
      id:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      place_name:
        type: string
      privacy:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
info:
  contact: {}
  description: This is a sample server celler server.
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
...
  /v1/memories:
    get:
      consumes:
      - application/json
      description: Get a list of memories.
      parameters:
      - description: Page number
        in: query
        minimum: 1
        name: page
        type: integer
      - description: Number of items per page
        in: query
        minimum: 1
        name: limit
        type: integer
      - description: Search by title, description, or tags
        in: query
        name: search_term
        type: string
      - description: Filter by multiple tags (comma-separated)
        in: query
        name: tags
        type: string
      - description: Filter by start date (YYYY-MM-DD)
        in: query
        name: start_date
        type: string
      - description: Filter by end date (YYYY-MM-DD)
        in: query
        name: end_date
        type: string
      - description: Filter by user ID
        in: query
        name: user_id
        type: string
      - description: Filter by title
        in: query
        name: title
        type: string
      - description: Filter by description
        in: query
        name: description
        type: string
      - description: Filter by latitude
        in: query
        name: latitude
        type: number
      - description: Filter by longitude
        in: query
        name: longitude
        type: number
      - description: Filter by place name
        in: query
        name: place_name
        type: string
      - description: Filter by privacy setting
        in: query
        name: privacy
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/memory.GetAllMemoriesResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: List Memories
      tags:
      - Memories
    post:
      consumes:
      - application/json
      description: Create a new memory with the provided details.
      parameters:
      - description: Memory details
        in: body
        name: memory
        required: true
        schema:
          $ref: '#/definitions/github_com_time_capsule_Api-Gateway-Timecapsule_models.CreateMemoryModel'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create a new Memory
      tags:
      - Memories
  /v1/memories/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a memory by its ID.
      parameters:
      - description: Memory ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete Memory
      tags:
      - Memories
    get:
      consumes:
      - application/json
      description: Get details of a memory by its ID.
      parameters:
      - description: Memory ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/memory.Memory'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get Memory by ID
      tags:
      - Memories
    patch:
      consumes:
      - application/json
      description: Partially update an existing memory.
      parameters:
      - description: Memory ID
        in: path
        name: id
        required: true
        type: string
      - description: Patched memory data
        in: body
        name: memory
        required: true
        schema:
          $ref: '#/definitions/github_com_time_capsule_Api-Gateway-Timecapsule_models.PatchMemoryModel'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Partially update Memory
      tags:
      - Memories
    put:
      consumes:
      - application/json
      description: Update an existing memory.
      parameters:
      - description: Memory ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated memory data
        in: body
        name: memory
        required: true
        schema:
          $ref: '#/definitions/github_com_time_capsule_Api-Gateway-Timecapsule_models.UpdateMemoryModel'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update Memory
      tags:
      - Memories
...
securityDefinitions:
  ApiKeyAuth:
    description: Description for what is this security definition being used
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"